Added tmt modifications for msfragger and mascot
